refactor(modules/commands): dir contents restructure
