+++ src/modules/commands/structures/handlers/builders/contextMenu.ts
@@ -1,0 +1,8 @@
+import { ContextMenuCommandBuilder } from "discord.js";
+import { AresBaseApplicationCommandData } from "../base";
+import { Mixin } from "ts-mixer";
+
+export class AresContextMenuCommandBuilder extends Mixin(
+  AresBaseApplicationCommandData,
+  ContextMenuCommandBuilder
+) {}
