+++ src/modules/commands/structures/handlers/builders/index.ts
@@ -1,0 +1,2 @@
+export * from "./chatInput";
+export * from "./contextMenu";
