+++ src/modules/commands/structures/handlers/index.ts
@@ -1,0 +1,2 @@
+export * from "./base";
+export * from "./chatInput";
