+++ src/utils/typeguards.ts
@@ -3,8 +3,8 @@ import {
   LogMessagesCodes,
   LogScopes,
 } from "../ts/enums";
-import { AresApplicationCommandType } from "../ts/types";
-import { AresChatInputCommand } from "../modules/commands/chatInputCommand";
+import { AresChatInputCommand } from "../modules/commands/structures/handlers";
+import { AresApplicationCommand } from "../ts/types";
 import AresEventHandler from "../modules/events/handler";
 import { Locale } from "discord.js";
 
@@ -29,10 +29,10 @@ export const isAresEventHandler = (
 
 export const isAresApplicationCommandType = (
   command: unknown
-): command is AresApplicationCommandType => {
+): command is AresApplicationCommand => {
   if (!command || command.constructor === Array) return false;
-  if (command instanceof AresChatInputCommand) return true;
-  return false;
+  if (!(command instanceof AresChatInputCommand)) return false;
+  return true;
 };
 
 export const isLocale = (str: unknown): str is Locale =>
